Add option to build forked wxWidgets libraries
This is for internal use only as it requires a forked version of wxWidgets and a
private version of the CMakeLists.txt file needed to build that version. This
allows for testing against current wxWidgets sources. The default is still to
build and use the libraries from the wxSnapshot sub-module.